requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib
bcrypt
python-jose
python-multipart
deta
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib